Move structs to standalone header
